+++ src/nodes/SetLiteral.ts
@@ -104,7 +104,7 @@ export default class SetLiteral extends CompositeLiteral {
     }
 
     getPurpose() {
-        return Purpose.Hidden;
+        return Purpose.Maps;
     }
 
     getAffiliatedType(): BasisTypeName | undefined {